% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-arima_boost.R
\name{sarima_lightgbm_fit_impl}
\alias{sarima_lightgbm_fit_impl}
\title{Bridge ARIMA-LightGBM Modeling function}
\usage{
sarima_lightgbm_fit_impl(
  x,
  y,
  period = "auto",
  p = 0,
  d = 0,
  q = 0,
  P = 0,
  D = 0,
  Q = 0,
  include.mean = TRUE,
  include.drift = FALSE,
  include.constant,
  lambda = model$lambda,
  biasadj = FALSE,
  method = c("CSS-ML", "ML", "CSS"),
  model = NULL,
  max_depth = 17,
  learning_rate = 0.1,
  num_iterations = 10,
  min_data_in_leaf = 20,
  min_gain_to_split = 0,
  bagging_fraction = 1,
  feature_fraction = 1,
  ...
)
}
\arguments{
\item{x}{A dataframe of xreg (exogenous regressors)}

\item{y}{A numeric vector of values to fit}

\item{period}{A seasonal frequency. Uses "auto" by default. A character phrase
of "auto" or time-based phrase of "2 weeks" can be used if a date or date-time variable is provided.}

\item{p}{The order of the non-seasonal auto-regressive (AR) terms.}

\item{d}{The order of integration for non-seasonal differencing.}

\item{q}{The order of the non-seasonal moving average (MA) terms.}

\item{P}{The order of the seasonal auto-regressive (SAR) terms.}

\item{D}{The order of integration for seasonal differencing.}

\item{Q}{The order of the seasonal moving average (SMA) terms.}

\item{include.mean}{Should the ARIMA model include a mean term? The default
is \code{TRUE} for undifferenced series, \code{FALSE} for differenced ones
(where a mean would not affect the fit nor predictions).}

\item{include.drift}{Should the ARIMA model include a linear drift term?
(i.e., a linear regression with ARIMA errors is fitted.)  The default is
\code{FALSE}.}

\item{include.constant}{If \code{TRUE}, then \code{include.mean} is set to
be \code{TRUE} for undifferenced series and \code{include.drift} is set to
be \code{TRUE} for differenced series. Note that if there is more than one
difference taken, no constant is included regardless of the value of this
argument. This is deliberate as otherwise quadratic and higher order
polynomial trends would be induced.}

\item{lambda}{Box-Cox transformation parameter. If \code{lambda="auto"},
then a transformation is automatically selected using \code{BoxCox.lambda}.
The transformation is ignored if NULL. Otherwise,
data transformed before model is estimated.}

\item{biasadj}{Use adjusted back-transformed mean for Box-Cox
transformations. If transformed data is used to produce forecasts and fitted values,
a regular back transformation will result in median forecasts. If biasadj is TRUE,
an adjustment will be made to produce mean forecasts and fitted values.}

\item{method}{Fitting method: maximum likelihood or minimize conditional
sum-of-squares. The default (unless there are missing values) is to use
conditional-sum-of-squares to find starting values, then maximum likelihood.}

\item{model}{Output from a previous call to \code{Arima}. If model is
passed, this same model is fitted to \code{y} without re-estimating any
parameters.}

\item{max_depth}{The maximum depth of the tree (i.e. number of splits).}

\item{learning_rate}{The rate at which the boosting algorithm adapts from iteration-to-iteration.}

\item{num_iterations}{The number of trees contained in the ensemble.}

\item{min_data_in_leaf}{The minimum number of data points in a node that is required for the node to be split further.}

\item{min_gain_to_split}{The reduction in the loss function required to split further.}

\item{bagging_fraction}{The amount of data exposed to the fitting routine.}

\item{feature_fraction}{The number of predictors that will be randomly sampled at each split when creating the tree models.}

\item{...}{Additional arguments passed to \code{lightgbm::lgb.train}}
}
\description{
Bridge ARIMA-LightGBM Modeling function
}
